kurs/contract_project/requirements.txt
python-telegram-bot==20.1
django==4.2
python-docx==0.8.11
nest-asyncio==1.5.6
aiofiles==24.1.0